changed datatypes

+++ src/main/java/com/anagha/Division.java
@@ -1,7 +1,7 @@
 package com.anagha;
 public class Division
 {
-    public float div(int a,int b) {
+    public float div(float a,float b) {
         float c= a/b;
         return c;
     }
